[MPOM-289] Use properties for all plugin versions
Alternate implementation to PR #155 to demonstrate my alternative
proposal with the naming convention of `version.artifactId` for
properties. The properties are sorted in the properties section.

This also includes individual properties for the plugins that use
`${maven.plugin.tools.version}` and `${surefire.version}`, so they can
be set individually, but will default to using the current properties
that set the version for multiple plugins at once.

This also fixes the M2Eclipse profile, `only-eclipse`, to remove the
upper bound on the version for maven-remote-resources-plugin, because
the version specified in this POM has already exceeded the that upper
bound, and the profile would not have been working correctly.